+++ libautomated.sh
@@ -263,7 +263,10 @@ to_file () {
         (
             set +e
 
-            diff -duaN "$target_path" - | tee >(printable_only | text_block "$1" | to_debug BRIGHT_BLACK >/dev/null) | patch -s -p0 "$target_path" >/dev/null
+            # TODO: AUT-125 Wait for process substitution subshell using { tee >(...); wait $!; }
+            # to prevent the script from moving onto the next command after the pipeline has
+            # completed, but the subshell is still sending output in background.
+            diff -duaN "$target_path" - | tee >(printable_only | text_block "$target_path" | to_debug BRIGHT_BLACK >/dev/null) | patch -s -p0 "$target_path" >/dev/null
 
             declare -a exit_codes=("${PIPESTATUS[@]}")
 
@@ -282,7 +285,9 @@ to_file () {
             fi
         )
     else
-        tee >(printable_only | text_block "$1" | to_debug BRIGHT_BLACK >/dev/null) >"$target_path"
+        # TODO: AUT-126 Write files atomically.
+        # shellcheck disable=SC2094
+        tee >(printable_only | text_block "$target_path" | to_debug BRIGHT_BLACK >/dev/null) >"$target_path"
     fi
 
     mtime_after=$(set -e; file_mtime "$target_path")
@@ -584,6 +589,7 @@ target_as_vars () {
 }
 
 
+# TODO: AUT-127 IPv6 addresses must be enclosed in square brackets.
 target_address_only () {
     declare target="$1"
     declare username address port
@@ -675,9 +681,19 @@ stdin_as_function () {
 
     printf 'drop_%s_body () {\n' "$file_id_hash"
 
-    stdin_as_code
+    declare code_sha256
+
+    # Waiting for process substitutions using wait $! is not supported in Bash 4.3
+    # or earlier. However, this function is mostly executed on locally on
+    # the controlling workstation, which is expected to have much more recent
+    # Bash version.
+    {
+        code_sha256=$(set -e; stdin_as_code | { tee >(cat >&${stdout}); wait $!; } | sha256sum | cut -f 1 -d ' ')
+    } {stdout}>&1
 
     printf '}\n'
+
+    printf 'AUTOMATED_DROP_%s_SHA256=%s\n' "${file_id_hash^^}" "$code_sha256"
 
     printf 'log_debug %q\n' "shipped STDIN as the file id ${file_id}"
 }
@@ -705,9 +721,19 @@ file_as_function () {
 
     printf 'drop_%s_body () {\n' "$file_id_hash"
 
-    stdin_as_code <"$src"
+    declare code_sha256
+
+    # Waiting for process substitutions using wait $! is not supported in Bash 4.3
+    # or earlier. However, this function is mostly executed on locally on
+    # the controlling workstation, which is expected to have much more recent
+    # Bash version.
+    {
+        code_sha256=$(set -e; stdin_as_code <"$src" | { tee >(cat >&${stdout}); wait $!; } | sha256sum | cut -f 1 -d ' ')
+    } {stdout}>&1
 
     printf '}\n'
+
+    printf 'AUTOMATED_DROP_%s_SHA256=%s\n' "${file_id_hash^^}" "$code_sha256"
 
     printf 'log_debug %q\n' "shipped ${src} as the file id ${file_id}"
 }
@@ -750,7 +776,7 @@ drop () {
     declare file_id="$1"
     declare dst="${2:-}"
 
-    declare mode file_id_hash mode_var
+    declare mode owner file_id_hash mode_var
 
     file_id_hash=$(md5 <<< "$file_id")
 
@@ -764,11 +790,38 @@ drop () {
             mode="${3:-}"
         fi
 
-        "drop_${file_id_hash}_body" "${file_id}" >"$dst"
+        owner="${4:-}"
+
+        # TODO: AUT-128 Support writing to pipes.
+
+        if ! [[ -e "$dst" ]]; then
+            touch "$dst"
+        fi
 
         if [[ -n "${mode:-}" ]]; then
             chmod "$mode" "$dst"
         fi
+
+        if [[ -n "${owner:-}" ]]; then
+            chown "$owner" "$dst"
+        fi
+
+        declare code_sha256_var
+        code_sha256_var="AUTOMATED_DROP_${file_id_hash^^}_SHA256"
+
+        if [[ -n "${!code_sha256_var:-}" ]]; then
+            log_debug "Checking if ${dst} needs updating"
+            declare code_sha256
+            code_sha256=$(stdin_as_code < "$dst" | sha256sum | cut -f 1 -d ' ')
+            if [[ "$code_sha256" == "${!code_sha256_var}" ]]; then
+                log_debug "${dst} is already up-to-date"
+                return 0
+            fi
+        fi
+
+        log_debug "Writing ${dst}"
+        # TODO: AUT-126 Write files atomically.
+        "drop_${file_id_hash}_body" "${file_id}" >"$dst"
     else
         "drop_${file_id_hash}_body" "$file_id"
     fi