+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2023-08-28
+### Added
+- Support for tmux 3.2 and later.
+- Singleton mode support.
+
+### Changed
+- `drop()` will ignore up-to-date target files.
+- Testing improvements.
+
+### Fixed
+- Check if target path exists before using diff mode in `to_file()`.
+
 ## [0.3.4] - 2022-04-10
 ### Added
 - `stdin_as_function()`.